style: 修复代码格式并添加 topic_builder.go 到 git
- 修复 envelope.go 注释格式
- 添加 topic_builder.go 到版本控制（之前被 gitignore）
